api/requirements-build.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-build.in
#
amqp==5.0.5
    # via kombu
asgiref==3.3.1
    # via django
billiard==3.6.3.0
    # via celery
celery==5.0.5
    # via -r requirements-build.in
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django-environ==0.4.5
    # via -r requirements-build.in
django-filter==2.4.0
    # via -r requirements-build.in
django==3.1.7
    # via
    #   -r requirements-build.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==4.6.0
    # via -r requirements-build.in
djangorestframework==3.12.4
    # via
    #   -r requirements-build.in
    #   djangorestframework-simplejwt
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements-build.in
gunicorn==20.1.0
    # via -r requirements-build.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
kombu==5.0.2
    # via celery
markupsafe==1.1.1
    # via jinja2
packaging==20.9
    # via drf-yasg
prompt-toolkit==3.0.18
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements-build.in
pyjwt==2.0.1
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via
    #   -r requirements-build.in
    #   celery
    #   django
redis==3.5.3
    # via -r requirements-build.in
requests==2.25.1
    # via coreapi
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.17.0
    # via drf-yasg
sentry-sdk==1.0.0
    # via -r requirements-build.in
six==1.15.0
    # via click-repl
sqlparse==0.4.1
    # via django
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.4
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
